some refine
