+++ pkg/infoschema/builder.go
@@ -105,8 +105,7 @@ func (b *Builder) ApplyDiff(m *meta.Meta, diff *model.SchemaDiff) ([]int64, erro
 }
 
 func (b *Builder) applyCreateTables(m *meta.Meta, diff *model.SchemaDiff) ([]int64, error) {
-	tblIDs := make([]int64, 0, len(diff.AffectedOpts))
-	return b.applyAffectedOpts(m, tblIDs, diff)
+	return b.applyAffectedOpts(m, make([]int64, 0, len(diff.AffectedOpts)), diff)
 }
 
 func (b *Builder) applyTruncateTableOrPartition(m *meta.Meta, diff *model.SchemaDiff) ([]int64, error) {
@@ -167,6 +166,7 @@ func (b *Builder) applyReorganizePartition(m *meta.Meta, diff *model.SchemaDiff)
 	return tblIDs, nil
 }
 
+// ywq todo the most complicate one
 func (b *Builder) applyExchangeTablePartition(m *meta.Meta, diff *model.SchemaDiff) ([]int64, error) {
 	// It is not in StatePublic.
 	if diff.OldTableID == diff.TableID && diff.OldSchemaID == diff.SchemaID {
@@ -473,78 +473,6 @@ func appendAffectedIDs(affected []int64, tblInfo *model.TableInfo) []int64 {
 	return affected
 }
 
-// copySortedTables copies sortedTables for old table and new table for later modification.
-func (b *Builder) copySortedTables(oldTableID, newTableID int64) {
-	if tableIDIsValid(oldTableID) {
-		b.copySortedTablesBucket(tableBucketIdx(oldTableID))
-	}
-	if tableIDIsValid(newTableID) && newTableID != oldTableID {
-		b.copySortedTablesBucket(tableBucketIdx(newTableID))
-	}
-}
-
-func (b *Builder) applyCreateOrAlterResourceGroup(m *meta.Meta, diff *model.SchemaDiff) error {
-	group, err := m.GetResourceGroup(diff.SchemaID)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	if group == nil {
-		return ErrResourceGroupNotExists.GenWithStackByArgs(fmt.Sprintf("(Group ID %d)", diff.SchemaID))
-	}
-	// TODO: need mark updated?
-	b.infoSchema.setResourceGroup(group)
-	return nil
-}
-
-func (b *Builder) applyDropResourceGroup(m *meta.Meta, diff *model.SchemaDiff) []int64 {
-	group, ok := b.infoSchema.ResourceGroupByID(diff.SchemaID)
-	if !ok {
-		return nil
-	}
-	b.infoSchema.deleteResourceGroup(group.Name.L)
-	// TODO: return the related information.
-	return []int64{}
-}
-
-func (b *Builder) applyCreatePolicy(m *meta.Meta, diff *model.SchemaDiff) error {
-	po, err := m.GetPolicy(diff.SchemaID)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	if po == nil {
-		return ErrPlacementPolicyNotExists.GenWithStackByArgs(
-			fmt.Sprintf("(Policy ID %d)", diff.SchemaID),
-		)
-	}
-
-	if _, ok := b.infoSchema.PolicyByID(po.ID); ok {
-		// if old policy with the same id exists, it means replace,
-		// so the tables referring this policy's bundle should be updated
-		b.markBundlesReferPolicyShouldUpdate(po.ID)
-	}
-
-	b.infoSchema.setPolicy(po)
-	return nil
-}
-
-func (b *Builder) applyAlterPolicy(m *meta.Meta, diff *model.SchemaDiff) ([]int64, error) {
-	po, err := m.GetPolicy(diff.SchemaID)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-
-	if po == nil {
-		return nil, ErrPlacementPolicyNotExists.GenWithStackByArgs(
-			fmt.Sprintf("(Policy ID %d)", diff.SchemaID),
-		)
-	}
-
-	b.infoSchema.setPolicy(po)
-	b.markBundlesReferPolicyShouldUpdate(po.ID)
-	// TODO: return the policy related table ids
-	return []int64{}, nil
-}
-
 func (b *Builder) applyCreateSchema(m *meta.Meta, diff *model.SchemaDiff) error {
 	di, err := m.GetDatabase(diff.SchemaID)
 	if err != nil {
@@ -604,16 +532,6 @@ func (b *Builder) applyModifySchemaDefaultPlacement(m *meta.Meta, diff *model.Sc
 	}
 }
 
-func (b *Builder) applyDropPolicy(PolicyID int64) []int64 {
-	po, ok := b.infoSchema.PolicyByID(PolicyID)
-	if !ok {
-		return nil
-	}
-	b.infoSchema.deletePolicy(po.Name.L)
-	// TODO: return the policy related table ids
-	return []int64{}
-}
-
 func (b *Builder) applyDropSchema(schemaID int64) []int64 {
 	di, ok := b.infoSchema.SchemaByID(schemaID)
 	if !ok {
@@ -657,6 +575,16 @@ func (b *Builder) applyRecoverSchema(m *meta.Meta, diff *model.SchemaDiff) ([]in
 		tables: make(map[string]table.Table, len(diff.AffectedOpts)),
 	}
 	return b.applyCreateTables(m, diff)
+}
+
+// copySortedTables copies sortedTables for old table and new table for later modification.
+func (b *Builder) copySortedTables(oldTableID, newTableID int64) {
+	if tableIDIsValid(oldTableID) {
+		b.copySortedTablesBucket(tableBucketIdx(oldTableID))
+	}
+	if tableIDIsValid(newTableID) && newTableID != oldTableID {
+		b.copySortedTablesBucket(tableBucketIdx(newTableID))
+	}
 }
 
 func (b *Builder) copySortedTablesBucket(bucketIdx int) {
@@ -889,46 +817,6 @@ func (b *Builder) copySchemasMap(oldIS *infoSchema) {
 	}
 }
 
-func (b *Builder) copyBundlesMap(oldIS *infoSchema) {
-	b.infoSchema.ruleBundleMap = make(map[int64]*placement.Bundle)
-	for id, v := range oldIS.ruleBundleMap {
-		b.infoSchema.ruleBundleMap[id] = v
-	}
-}
-
-func (b *Builder) copyPoliciesMap(oldIS *infoSchema) {
-	is := b.infoSchema
-	for _, v := range oldIS.AllPlacementPolicies() {
-		is.policyMap[v.Name.L] = v
-	}
-}
-
-func (b *Builder) copyResourceGroupMap(oldIS *infoSchema) {
-	is := b.infoSchema
-	for _, v := range oldIS.AllResourceGroups() {
-		is.resourceGroupMap[v.Name.L] = v
-	}
-}
-
-func (b *Builder) copyTemporaryTableIDsMap(oldIS *infoSchema) {
-	is := b.infoSchema
-	if len(oldIS.temporaryTableIDs) == 0 {
-		is.temporaryTableIDs = nil
-		return
-	}
-
-	is.temporaryTableIDs = make(map[int64]struct{})
-	for tblID := range oldIS.temporaryTableIDs {
-		is.temporaryTableIDs[tblID] = struct{}{}
-	}
-}
-
-func (b *Builder) copyReferredForeignKeyMap(oldIS *infoSchema) {
-	for k, v := range oldIS.referredForeignKeyMap {
-		b.infoSchema.referredForeignKeyMap[k] = v
-	}
-}
-
 // getSchemaAndCopyIfNecessary creates a new schemaTables instance when a table in the database has changed.
 // It also does modifications on the new one because old schemaTables must be read-only.
 // And it will only copy the changed database once in the lifespan of the Builder.
@@ -948,26 +836,6 @@ func (b *Builder) getSchemaAndCopyIfNecessary(dbName string) *model.DBInfo {
 		return newSchemaTables.dbInfo
 	}
 	return b.infoSchema.schemaMap[dbName].dbInfo
-}
-
-func (b *Builder) initMisc(dbInfos []*model.DBInfo, policies []*model.PolicyInfo, resourceGroups []*model.ResourceGroupInfo) {
-	info := b.infoSchema
-	// build the policies.
-	for _, policy := range policies {
-		info.setPolicy(policy)
-	}
-
-	// build the groups.
-	for _, group := range resourceGroups {
-		info.setResourceGroup(group)
-	}
-
-	// Maintain foreign key reference information.
-	for _, di := range dbInfos {
-		for _, t := range di.Tables {
-			b.infoSchema.addReferredForeignKeys(di.Name, t)
-		}
-	}
 }
 
 func (b *Builder) initVirtualTables(schemaVersion int64) error {
@@ -1088,13 +956,6 @@ func (b *Builder) addTable(schemaVersion int64, di *model.DBInfo, tblInfo *model
 		sortedTbls := b.infoSchema.sortedTablesBuckets[tableBucketIdx(tblInfo.ID)]
 		b.infoSchema.sortedTablesBuckets[tableBucketIdx(tblInfo.ID)] = append(sortedTbls, tbl)
 	}
-}
-
-func (b *Builder) addTemporaryTable(tblID int64) {
-	if b.infoSchema.temporaryTableIDs == nil {
-		b.infoSchema.temporaryTableIDs = make(map[int64]struct{})
-	}
-	b.infoSchema.temporaryTableIDs[tblID] = struct{}{}
 }
 
 type virtualTableDriver struct {
